Add classifier model handling

+++ build_all_images.sh
@@ -3,6 +3,7 @@
 BUILD_DIR=image_build
 COLEXT_DIR="domainradar-colext"
 INPUT_DIR="domainradar-input"
+CLF_DIR="domainradar-clf"
 
 cd dockerfiles || exit 1
 
@@ -27,14 +28,32 @@ else
   git clone git@github.com:nesfit/domainradar-colext.git "$COLEXT_DIR"
 fi
 
+if [ -d "$CLF_DIR" ]; then
+  echo "Pulling latest domainradar-clf"
+  cd "$CLF_DIR" || exit 1
+  git pull
+  cd ..
+else
+  git clone git@github.com:nesfit/domainradar-clf.git "$CLF_DIR"
+fi
+
 cd ..
 
 echo "Building the prefilter image"
 docker build -f prefilter.Dockerfile -t domrad/prefilter .
 
 echo "Building the pipeline images"
-cd "$BUILD_DIR/$COLEXT_DIR" || (rm -rf image_build && exit 1)
+cd "$BUILD_DIR/$COLEXT_DIR" || exit 1
 ./build_docker_images.sh
 
 cd ../..
-rm -rf image_build
+
+echo "Copying models"
+rm -rf ../clf_models
+mkdir -p ../clf_models
+cp -r "$BUILD_DIR/$CLF_DIR/classifiers/models" ../clf_models/models
+cp -r "$BUILD_DIR/$CLF_DIR/classifiers/boundaries" ../clf_models/boundaries
+
+if [ "${DELETE_CLONED:-0}" = "1" ]; then
+  rm -rf image_build
+fi
